extracted MainInformationView view setup login into a extension
